server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.2
bcrypt==4.0.1
pydantic==2.5.1
groq==0.4.1
PyPDF2==3.0.1
gunicorn==21.2.0